+++ src/Header/Request/Range.php
@@ -41,7 +41,8 @@ final class Range implements Header
     public static function fromValue(string $value): self
     {
         $regEx = '{^' . Rfc7233::RANGE . '$}';
-        if (utf8_decode($value) !== $value || preg_match($regEx, $value) !== 1) {
+        $plainValue = mb_convert_encoding($value, 'ISO-8859-1', 'UTF-8');
+        if ($plainValue !== $value || preg_match($regEx, $value) !== 1) {
             throw new InvalidHeader('Invalid header value: ' . $value);
         }