+++ src/Header/Value/CacheControl/Directive.php
@@ -28,7 +28,8 @@ abstract class Directive
     public static function fromString(string $directive): self
     {
         $regEx = '{^' . Rfc7234::CACHE_DIRECTIVE_CAPTURE . '$}';
-        if (utf8_decode($directive) !== $directive || preg_match($regEx, $directive, $matches) !== 1) {
+        $plainDirective = mb_convert_encoding($directive, 'ISO-8859-1', 'UTF-8');
+        if ($plainDirective !== $directive || preg_match($regEx, $directive, $matches) !== 1) {
             throw new InvalidArgumentException('Invalid directive: ' . $directive);
         }