+++ src/Header/Value/CacheControl/GeneralDirective.php
@@ -41,7 +41,8 @@ final class GeneralDirective extends Directive
     public static function fromString(string $directive): self
     {
         $regEx = '{^' . Rfc7234::CACHE_DIRECTIVE_CAPTURE . '$}';
-        if (utf8_decode($directive) !== $directive || preg_match($regEx, $directive, $matches) !== 1) {
+        $plainDirective = mb_convert_encoding($directive, 'ISO-8859-1', 'UTF-8');
+        if ($plainDirective !== $directive || preg_match($regEx, $directive, $matches) !== 1) {
             throw new InvalidArgumentException('Invalid directive: ' . $directive);
         }
 
@@ -75,7 +76,8 @@ final class GeneralDirective extends Directive
      */
     public function setName(string $name): void
     {
-        if (utf8_decode($name) !== $name || preg_match('{^' . Rfc7230::TOKEN . '$}', $name) !== 1) {
+        $plainName = mb_convert_encoding($name, 'ISO-8859-1', 'UTF-8');
+        if ($plainName !== $name || preg_match('{^' . Rfc7230::TOKEN . '$}', $name) !== 1) {
             throw new InvalidArgumentException('Invalid name: ' . $name);
         }
 
@@ -109,7 +111,8 @@ final class GeneralDirective extends Directive
             $encodedValue = self::encodeValue($value);
 
             $regEx = '{^(?:' . Rfc7230::TOKEN . '|' . Rfc7230::QUOTED_STRING . ')$}';
-            if (utf8_decode($encodedValue) !== $encodedValue || preg_match($regEx, $encodedValue) !== 1) {
+            $plainValue = mb_convert_encoding($encodedValue, 'ISO-8859-1', 'UTF-8');
+            if ($plainValue !== $encodedValue || preg_match($regEx, $encodedValue) !== 1) {
                 throw new InvalidArgumentException('Invalid value: ' . $value);
             }
         }