+++ src/Header/Value/CacheControl/IntegerDirective.php
@@ -46,7 +46,8 @@ final class IntegerDirective extends Directive
     public static function fromString(string $directive): self
     {
         $regEx = '{^' . Rfc7234::CACHE_DIRECTIVE_CAPTURE . '$}';
-        if (utf8_decode($directive) !== $directive || preg_match($regEx, $directive, $matches) !== 1) {
+        $plainDirective = mb_convert_encoding($directive, 'ISO-8859-1', 'UTF-8');
+        if ($plainDirective !== $directive || preg_match($regEx, $directive, $matches) !== 1) {
             throw new InvalidArgumentException('Invalid directive: ' . $directive);
         }
 
@@ -76,7 +77,8 @@ final class IntegerDirective extends Directive
      */
     public function setName(string $name): void
     {
-        if (utf8_decode($name) !== $name || preg_match('{^' . Rfc7230::TOKEN . '$}', $name) !== 1) {
+        $plainName = mb_convert_encoding($name, 'ISO-8859-1', 'UTF-8');
+        if ($plainName !== $name || preg_match('{^' . Rfc7230::TOKEN . '$}', $name) !== 1) {
             throw new InvalidArgumentException('Invalid name: ' . $name);
         }
 
@@ -106,7 +108,8 @@ final class IntegerDirective extends Directive
     public function setValue(string $value): void
     {
         $regEx = '{^' . Rfc7234::DELTA_SECONDS . '$}';
-        if (utf8_decode($value) !== $value || preg_match($regEx, $value) !== 1) {
+        $plainValue = mb_convert_encoding($value, 'ISO-8859-1', 'UTF-8');
+        if ($plainValue !== $value || preg_match($regEx, $value) !== 1) {
             throw new InvalidArgumentException('Invalid value: ' . $value);
         }